requirements.txt
streamlit
openai
moviepy
google-genai
google-cloud-videointelligence
google-cloud-vision
boto3
openai-whisper
google-generativeai
google-cloud-videointelligence
google-cloud-vision
pillow
# opencv-python
opencv-python-headless
google-cloud-speech
firebase-admin
pinecone
librosa
speechbrain
azure-cognitiveservices-speech
streamlit-extras
flask
celery